Added give exp component
Added an actor component to make giving the player exp easier, load this
component on any enemy or chalange you want to give exp to the player
and call it's giveexp function when appropriate. there are two inputs.
one is a player index, juszt leave this at 1 for now, I made it so in
the future it will be easier if we want to incorperate multiplayer, the
next is the EXP you wish to give